backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
dotenv==0.9.9
fastapi==0.128.0
fonttools==4.61.1
frozenlist==1.8.0
h11==0.16.0
idna==3.11
inference-sdk==0.64.2
kiwisolver==1.4.9
marshmallow==3.26.2
matplotlib==3.10.8
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.5
opencv-python==4.10.0.84
packaging==25.0
pillow==11.3.0
propcache==0.4.1
py-cpuinfo==9.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.0
six==1.17.0
starlette==0.50.0
supervision==0.27.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
yarl==1.22.0
